Asset Upload Support
Upgrade the applause-reporter-common package so we can upload assets to the Applause Automation API.

We needed to transition our reporter to multiple services to gain access to more hooks that webdriverio provides. This also means that we now require two services to exeucte tests against the automation api. This is because services can only be classified as either launcher services, or worker services. The RunService is considered a launcher service, while the result service is a worker service.

To better handle logging for results, we built in the ability to inject a custom winston logger into the services. If one is not provided, we construct a default logger that is used for each component of the service.